Terrain, ISO sphere fog, Cam pitch, radar scrolling
Started terrain, still seams and artifacts.

Iso Fog, by arainging 'particals' around the player in a Isometric pattern, the player is guarenteed to always be sourounded with minimal excess, a custom shaer can fade from the world to the fog.

Radars can scroll, rings and blips move. Click to drive does not apply this transorm yet and will behave as if no scrolling were in effect. (No user method to scroll yet)